Significant documentation improvements and demos
