+++ app/viewer/components.js
@@ -2,12 +2,12 @@
 
 //import the css used in this app and config items for optimized build
 import 'node_modules/openlayers/dist/ol.css!';
-import '../../node_modules/bootstrap/dist/css/bootstrap.min.css!';
-import '../../node_modules/bootstrap/dist/css/bootstrap-theme.min.css!';
-import '../../node_modules/font-awesome/css/font-awesome.min.css';
+import 'bootstrap/dist/css/bootstrap.min.css!';
+import 'bootstrap/dist/css/bootstrap-theme.min.css!';
+import 'font-awesome/css/font-awesome.min.css';
 import './css/viewer.css!';
 
-import 'bootstrap';
+import 'bootstrap/dist/js/bootstrap.min';
 
 //import components used by this app
 //TODO: these may be converted into <can-import> tags eventually
