+++ components/filter-widget/demo.html
@@ -1,4 +1,4 @@
-<script type="text/stache" id="demo" can-autorender>
+<script type="text/stache" id="demo-html" can-autorender>
   <h1>Filter Widget Demo</h1>
   <filter-widget />
 </script>
