+++ components/form-widget/demo.html
@@ -1,24 +1,66 @@
-<script type="text/stache" id="demo" can-autorender>
+<script type="text/stache" id="demo-html" can-autorender>
+  <p>Current form values: </p>
+  <ul>
+    <li>{{formObject.field1}}</li>
+    <li>{{formObject.field2}}</li>
+    <li>{{formObject.field3}}</li>
+    <li>{{formObject.field4}}</li>
+    <li>{{formObject.field5}}</li>
+  </ul>
+
   <h1>Form Widget Demo</h1>
-  <form-widget {form-object}="formObject" {fields}="fields" />
+  <form-widget {(form-object)}="formObject" {fields}="fields" (submit)="onSubmit" />
 </script>
 <script src="../../node_modules/steal/steal.js" main="@empty" id="demo-source">
   import 'bootstrap/dist/css/bootstrap.min.css!';
   import 'bootstrap/dist/css/bootstrap-theme.min.css!';
   import 'font-awesome/css/font-awesome.min.css';
+
+  //import our form component
   import 'components/form-widget/';
 
-  import can from 'can/view/';
+  //import our fields that we're using
+  import 'components/form-widget/field-components/text-field/';
+  import 'components/form-widget/field-components/select-field/';
+  import 'components/form-widget/field-components/file-field/';
+  import 'components/form-widget/field-components/json-field/';
+  import 'components/form-widget/field-components/date-field/';
 
-  var frag = can.view('demo', {
-    formObject: new can.Map({
-      field1: 'test-value',
-      field2: 'another value',
-      field3: 1
-    }),
+  //import canjs stuff
+  import can from 'can/view/';
+  import 'can/util/fixture/';
+
+  //this example uses fixtures to catch requests and simulate
+  //file upload experience
+  can.fixture('POST /upload', function(request) {
+    return {
+      uploads: ['fake_upload_' + request.data.entries().next().value[1].name]
+    };
+  });
+
+  can.fixture('DELETE /upload', function() {
+    return {
+      result: 'Success'
+    };
+  });
+
+  var template = can.Map.extend({
+    field1: 'test-value',
+    field2: 'another value',
+    field3: 1,
+    field4: '',
+    field5: ''
+  });
+
+  var frag = can.view('demo-html', {
+    onSubmit: function(){
+      alert('Form submitted! See the console for details');
+      console.log(arguments);
+    },
+    formObject: new template(),
     fields: ['field1', {
       name: 'field2',
-      alias: 'A better field name',
+      alias: 'A basic textarea field',
       type: 'text',
       properties: {
         textarea: true
@@ -31,10 +73,21 @@
         options: [{
           value: 1,
           label: 'Option 1'
-        },{
+        }, {
           value: 2,
           label: 'Option 2'
         }]
+      }
+    }, {
+      name: 'field4',
+      alias: 'A date field',
+      type: 'date'
+    }, {
+      name: 'field5',
+      alias: 'A file field',
+      type: 'file',
+      properties: {
+        url: '/upload'
       }
     }]
   }, {});
