+++ components/form-widget/field-components/date-field/date-field.md
@@ -1,0 +1,9 @@
+
+@module {can.Component} components/form-widget/field-components/date-field <date-field />
+@parent form-widget.fields
+
+@option {String} type This should be set to `date` for the form widget to use this template
+
+@description
+
+A date picker text box. Note: This field type has specific dependencies on the project *date-selector* which requires jquery, underscore, and moment.
