+++ components/form-widget/field-components/file-field/file-field.md
@@ -1,0 +1,12 @@
+@module {can.Component} components/form-widget/field-components/file-field <file-field />
+@parent form-widget.fields
+
+@option {String} type Should be set to `file` for the form to use this template
+@option {String} properties.url The url up post the file upload to
+@option {Boolean} properties.multiple Whether or not to allow multiple uploads
+@option {String} properties.accept The type of media to accept
+@link http://www.w3schools.com/tags/att_input_accept.asp Input Accept Types
+
+@description
+
+A file upload field. Once a file is selected, this component POST's the upload to a special url and returns the url result as text to the form.
