+++ components/form-widget/field-components/select-field/SelectFieldOption.md
@@ -1,0 +1,5 @@
+
+@typedef {Object} geocola.types.SelectOptionProperty SelectOptionProperty
+@parent geocola.types
+@option {String} value The value of the dropdown. This is converted to a string inside the option tag.
+@option {String} label The label to display in the select dropdown.
