+++ components/form-widget/field-components/select-field/select-field.stache
@@ -1,7 +1,7 @@
 <div class="form-group">
     <label for="{{name}}">{{alias}}</label>
     <select ($change)="onChange(%element.value)" id="{{name}}" class="form-control" name="{{name}}">
-      <option value="" {{^value}} selected {{/value}}></option>
+      <option value="" {{^value}} selected {{/value}}>{{placeholder}}</option>
       {{#each options}}
         <option value="{{value}}" {{#isSelected(value)}} selected {{/isSelected}}>{{label}}</option>
       {{/each}}
