--- components/form-widget/field-components/text-field.js
@@ -1,36 +1,0 @@
-import widgetModel from 'components/widget-model';
-import can from 'can';
-import template from './text-field.stache!';
-/**
- * @module form-widget.fields.text Text
- * @parent form-widget.fields
- * @body
- * A generic text box with a label. If the `type` property is not provided, `'text'` is the default.
- *
- * `type: 'text'`
- */
-
- /**
-  * @typedef {textFieldProperties} form-widget.types.textFieldProperties textFieldProperties
-  * @parent form-widget.types
-  * @option {String} type The type of textbox, this can be any html input type
-  * @option {Boolean} textarea Set to true if the input textbox should be a large textarea
-  * @link http://www.w3schools.com/html/html_form_input_types.asp HTML Input Types
-  */
-
-export let ViewModel = widgetModel.extend({
-  define: {
-    properties: {
-      Value: can.Map
-    }
-  },
-  onChange(element) {
-    this.dispatch('change', [element.value]);
-  }
-});
-
-can.Component.extend({
-  tag: 'text-field',
-  template: template,
-  viewModel: ViewModel
-});
