+++ components/form-widget/guides/FormFieldChange.md
@@ -1,5 +1,6 @@
 
  @typedef {can.Event} components/form-widget.ViewModel.events.formFieldChangeEvent fieldChange
- An event dispatched when a form field changes. The formObject is passed as an argument
  @parent components/form-widget.ViewModel.events
  @option {can.Map} formObject The formObject
+ @description
+ An event dispatched when a form field changes. The formObject is passed as an argument
