+++ components/form-widget/guides/FormSubmitEvent.md
@@ -1,4 +1,7 @@
+
 @typedef {can.Event} components/form-widget.ViewModel.events.formSubmit submit
 @parent components/form-widget.ViewModel.events
+@option {can.Map} formObject The form object that is saved
+@description
+
 An event dispatched when the form is submitted
-@option {can.Map | Object} formObject The form object that is saved
