+++ package.json
@@ -19,8 +19,8 @@
   },
   "dependencies": {
     "bootstrap": "^3.3.6",
-    "can": "^2.3.22",
-    "can-connect": "^0.5.1",
+    "can": "^2.3.23",
+    "can-connect": "^0.5.4",
     "date-selector": "^2.0.0",
     "font-awesome": "^4.5.0",
     "jquery": "^2.2.3",
